requirements.txt
streamlit
torch
torchvision
pillow
numpy
pandas
opencv-python
